Update puppetctl to 2019-modern
This catches up the repo to changes we made over the past 4-5 years.
shellcheck cleanup and exit code standardization, as well as parsing changes.
But the major introduction was 'operate' and 'nooperate' directives, which allow you to have puppet continue running but only in noop mode, for a particular time duration.